VBV-783 - Cannot auto-discover a network if there is another one with the same name in RETIRED state
The main problem was basically that we considered the discovered networks
unique by name, and after one was created no other network could have the
same name.

- Removed that restriction by making the network id part of the state self
link instead of the name since it won't be unique anymore (again, only for
discovered networks).
- Added the attribute "connected" to the network state and UI to know,
specially in case when you have multiple retired networks, the time when
Admiral "connected" to that network. It can't be retrieved from the Docker
API.
- Fixed a small issue caused with the delay of the network maintenance, it
affected just the discovered networks, that appeared with state UNKNOWN for
a while.
- Updated the ContainerToNetworkAffinityHostFilter since now we accept
multiple external networks with the same name.
- Removed the id of the network from the UI selection when editing a
template because discovered networks will be picked by name at runtime
independently of the name/id that the user picks in the template.

Change-Id: Ie3258be11ad87b02885dfc9651429a1c41a3ab17
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/3800
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Lazarin Lazarov <llazarov@vmware.com>