Fix sporadic failures in testRemoveApplicationWithScaledContainer
There is an issue when removing an application, if several containers share
the same container description, when removing those containers it is possible
that the container description is already removed. We are handling the case
when status code is not found, but it is also possible to get a
CancellationException, that should be handled in the same way.

Change-Id: Ia573f02000445c4702601600a142305e744bbc19
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/3822
Reviewed-by: Sergio Sanchez <sanchezs@vmware.com>
Bellevue-Verified: e_vcoauto_glob_1 <e_vcoauto_glob_1@vmware.com>
CS-Verified: e_vcoauto_glob_1 <e_vcoauto_glob_1@vmware.com>