Fix for Scale-out docker containers with on-demand network definition fails with "Network Not Found" error
- Updated the ContainerToNetworkAffinityHostFilter to be aware of the actual
network names when trying to filter hosts which contain such networks. The
problem was that for external networks the name is always the same (between
the network description and the network state), but for on-demand networks
the name changes between description and state... and that was causing
problems on scaling operations. Now the actual network name is in use when
available.
- Added a couple of unit tests to exercise the scaling scenario with
on-demand networks when the containers can be scaled and when they can't be
scaled because of port bindings conflicts.

Change-Id: Ic2dc422a657d3671a5b4ffb465d40fe0c44ba0ed
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/3836
Reviewed-by: Tony Georgiev <tgeorgiev@vmware.com>
Bellevue-Verified: e_vcoauto_glob_1 <e_vcoauto_glob_1@vmware.com>
CS-Verified: e_vcoauto_glob_1 <e_vcoauto_glob_1@vmware.com>