Bug 1762816: Application is removed after removing a container if there are external networks only
If a template with networks is provisioned, and all the networks are external
(no on-demand networks), when a container in this application is removed
(i.e. scaled in or removed), then the composite component is also removed,
but it should not.

The condition to remove the composite component is to check if the remaining
components are all external networks. However, when there are containers and
external networks only, the application was removed, but it this case, it should
be kept.

Bug Number: 1762816

Change-Id: Ia7293a1d4207ed8bf3f117047bb5db38dfbd5f01
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/4008
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Lazarin Lazarov <llazarov@vmware.com>