Fixing UI elements position regressions
there are 2 issue regressed recently from the 2 major style improvements
1 - There was no padding in the template details label and name.
The problem was that we were relying on an HTML template issue,
that there were empty spaces between the html elements, now that we minify
the templates, this was broken.

2. Network anchors were visible only on first container
Now the problem this was happening was because JSplumb was calculating
only the offset top and left positions of an element in order to place the anchor
elements. However we made a change recenlty that changed the components
to use css transform instead, and this was not taken into account.

Change-Id: I80bc855593b59abb4a7f57dd2288fce5ace298dd
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/4106
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Sergio Sanchez <sanchezs@vmware.com>