VBV-824 TEST: RequestBrokerServiceTest.testRequestLifeCycleWithContainerNetworkAndServiceAntiAffinityFilterFailureShouldCleanNetworks
The issue is that sometimes the ContainerToNetworkAffinityFilter filters out
the wrong host. In the scenario the test is checking both containers should be
placed on the same host according to the ContainerToNetworkAffinityFilter.
We get to the point where, for the first container, we choose a host out of
the available ones (we do this by turning a map to a list and picking an
element at a specific index). We have to do the same for the second container
and put that container on the same host. If the map is a HashMap, the order
of iteration is not guaranteed and we may end up choosing a different host. All
of this is accounted for and the maps are converted to TreeMaps, however, in
one place the HashMap is passed instead of the TreeMap and this causes the bug

Bug Number:
Documentation Notes:
PrecheckinTest URL: <required>
QA Notes:
Review URL: <required>
Reviewed by: <required>
Testing Done: ran the test many times

Change-Id: I88a893108d48d29d6c732dc91935816c90ec86a9
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/4108
CS-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Sergio Sanchez <sanchezs@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>