VBV-810: Scale-out docker containers with links fails when multiple hosts are available
When clustering a container in an application and other containers in the
application have links to this container, the current container must be
placed in the same host where other instances of this container are already
placed. If not, we can end up in a situation where we scale up a container
with links, and this container can have links to containers in different hosts,
which is not supported.

To avoid the issue, the new filter ClusterServiceLinkAffinityHostFilter will
place new instances of the linked container in the same host where the previous
instances are.

Change-Id: Iea61e99fc5efe619d894e231f5245082d14b62a9
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/4231
Bellevue-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Sergio Sanchez <sanchezs@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/4297