Fix for removing custom properties by a PATCH request
Sending a null value used to work in a single node where no JSON
serialization is done but fails with remote calls because null values
are excluded from the patch body. A custom serialization is now
configured when the PATCH body is a ResourceState object.

The fix here is local to admiral (a more generic solution is possible
in xenon).

Change-Id: I2cec9e022c2eef4095daace831d1943a8556dc8a
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/4227
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Rostislav Georgiev <rgeorgiev@vmware.com>