Fix for removing custom properties by a PATCH request
Sending a null value used to work in a single node where no JSON
serialization is done but fails with remote calls because null values
are excluded from the patch body. A custom serialization is now
configured when the PATCH body is a ResourceState object.

The fix here is local to admiral (a more generic solution is possible
in xenon).

Change-Id: I907f87a7b049bfc456cfdeb71783dfd67ac328e4
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/4252
CS-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Sergio Sanchez <sanchezs@vmware.com>