VBV-841 - Inconsistent data collection in cluster mode
The issue was reproducible only in vRA cluster. The issue was that
during the data collection before creating a container state we try to
get it and in case of 404 we create it. The problem is that in the
callback there is a check for exception and if there is exception the
status code is check. In vRA cluster (https) 404 is always returned but
not every time we have an exception in the callback.
- Changed the check for a container state before creating. Now first the
status code will be checked
- Did the same change on the other places where we had such order

Change-Id: I2ad5715817f37846f2b23381d0aa55337a6b3b8d
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/4264
Reviewed-by: Sergio Sanchez <sanchezs@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Igor Stoyanov <istoyanov@vmware.com>