Bug 1762816: Application is removed after removing a container if there are external networks only
If a template with networks is provisioned, and all the networks are external
(no on-demand networks), when a container in this application is removed
(i.e. scaled in or removed), then the composite component is also removed,
but it should not.

The condition to remove the composite component is to check if the remaining
components are all external networks. However, when there are containers and
external networks only, the application was removed, but it this case, it should
be kept.

Bug Number: 1762816

Change-Id: I1b274de97f241baa3d855bdfe51580d52422fda0
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/4296
CS-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Sergio Sanchez <sanchezs@vmware.com>