[VBV-764] Wrong number of containers in discovered network
Before the bump of the remote API to version 1.21 it was not possible to
guess the networks to which a discovered container was connected.
Updated the code to have the networks collected during a container
inspect. System-defined docker networks are ommited during collection.

Change-Id: Icb9a1fe1739a087ac22da52d70f92fabb568118f
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/4316
Reviewed-by: Sergio Sanchez <sanchezs@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>