Add test for default bootstrap services
The test validates that when a cluster is running
all of the hosts have the right bootstrap services created and are active.

Also added check for ServiceAlreadyStartedException in order to pass
the initial boot creation.

Revesed the check for equality of cert and existing credential
publickey to not protect against NPE

This change also re-enables the ManagementHostCluster tests
Change-Id: I997641fe7199d7e64a92afca7695e30b2f8516d2
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/4367
Reviewed-by: Igor Stoyanov <istoyanov@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>