VBV-862: Configure Xenon clustering on different port
- Add new parameter nodeGroupPublicUri. If it's set, start a new listener on
it. It must be used as peerList and advertised as publicUri. Also, if key /
certificate are provided, create an ssl context for the listener. Note that
a different authentication mechanism can be used for the replication traffic,
but for now we are reusing the same one.
- Add some validations for nodeGroupPublicUri: it's well formed and distinct
than uri + port.
- Add some unit tests. Pending some more unit tests in clustering.
- Additionally, one small fix in startSwaggerService: it must use getUri
instead of getPublicUri, because getPublicUri is the address to advertise
to peers (although the variable name can be a bit misleading).

Change-Id: I45589d6e72083f6d9241ea22789a0aace69bf9ae
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/4440
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Sergio Sanchez <sanchezs@vmware.com>