Fix failing RequestBrokerServiceTest and more
Fixing testRequestLifeCycleWithContainerNetworkAndPortBindingConflictOnScaleShouldFail.
When deploying two containers within the same Application, which expose
the same ports on the host, they were getting  deployed at the same time. Because of this,
the ExposedPortsHostFilter didn't know to distribute them on different
hosts.

Additionally, we were not setting the ports field of the initial
ContainerState we create. Since we wait for all ContainerStates to get
created (allocated) before calling the DockerAdapter, we didn't know about the
exposed ports of previously allocated containers until all containers were
placed.

Fixing this test uncovered a number of tests with bad container
description configurations + an issue with
ContainerToNetworkAffinityHostFilter, where a different host was chosen
for each container. This was fixed by moving the ExposedPortsHostFilter
down with the rest of the affinity filters.

Change-Id: Ifacd0f2177fca5acce72bf02a57dba2284d093ab
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/4549
Bellevue-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Sergio Sanchez <sanchezs@vmware.com>
Reviewed-by: Igor Stoyanov <istoyanov@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>