Trigger data collection more often after starting a node
Trigger the data collection every minute during the first 3 minutes.
This is needed because in cluster when a node is restarted the
certificates are not loaded on start but after 1 minute. In this case if
the DC is triggered before the certificates are loaded the existing
docker hosts will be marked as DISABLED. 
- Trigger the data collection more often in order to get the hosts in ON
state sooner.

Change-Id: I11e7aef7e3a8e1b428d7c720e7f464d188f40d33
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/4578
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Jose Dillet <jdillet@vmware.com>