Remove reconfiguring of the log formatters
This method has left after trying to format time as human readable string.
Now that's done in xenon and not necessary in here.
Apart from that the colored log formatter and the log formatter are used
exactly the opposite as in ServiceHost and not the ANSI codes are written
in xenon.log file and not in to the console.

Change-Id: I91933d554d1088a36cd98587a11b5458f673cb2f
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/4859
Reviewed-by: Elena Ivanova <eivanova@vmware.com>
Reviewed-by: Tony Georgiev <tgeorgiev@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>