UI: Fix openNetworks action in multiple views
In multiple container related views there is a "networks" icon that can be
clicked to show the networks that are related to a given container.
However, if there are more than one network, clicking the icon will
display an empty view, because the openNetworks method in these views is
not setting the search occurennce to "any". This change fixes that.

Change-Id: I378a10edbd9a26dce1d099150780b26b2702f0cb
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/4899
Reviewed-by: Iveta Ilieva <iilieva@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>