Commands will work with names also.
Everywhere the user had to provide ID for specific entity now it's
possible to provide the name of the entity as long as the name is
unique.

Still the priority of filtering is ID > Name. This means that with the
provided input the CLI first assume it is ID and it will look for
entities with this ID, in case it don't find entities with this ID it
will look for entities with name.

Change-Id: Idba18694febccb5cfa94ac02bc3a425ec1d3c4e3
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/4902
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Sergio Sanchez <sanchezs@vmware.com>