Fix related to https connections.
The CLI will not depend on Golang http client to handle pool of trusted
certs. Instead of this trusted certificates will be loaded and when
there is problematic one the CLI will check if it's contained in the
file with already trusted certs and if it is there it will turn off tls
verifictaion if it's not there it will prompt the user to accept it or
decline it.

Also there is small change for formatting the output when requests are
being listed.

Change-Id: Ie03cf984e391edab1ff7a4020f93f9f1ab6c602c
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/4930
Reviewed-by: Sergio Sanchez <sanchezs@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>