Assure deleting containerStates when deleting containers.
In performance job fails with several containter states left
after container removal request. The rason for this is a race
condition between the removal request and the data conllection.
The docker container is removed in the same time data collection
updates the state. This result with a containerState with missing
docker container. Later the data collection mark the state as
retired.

Change-Id: I5b90a57b36e9ea0874bde0a4b834e37a9ee788b1
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/5178
CS-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Aleksandar Angelov <aangelov@vmware.com>