VBV-856 - Add a proper authN for container shell proxy.
Step 1:
- Enabled SSL on the agent side so all the communication between Admiral
and the agent is over SSL.
- Current approach: Generate a self-signed certificate when the the agent is
started. And the certificate will be imported to Admiral when the system
container is discovered. Each agent will have a different certificate. One
of the CONS is that they will appear in the list of Certificates in the UI.

Next step:
- Enable SSL client side authentication so that *only* Admiral can
communicate with the agent.

Change-Id: Iafddb320bc5276c893ffd00e56d78e0d3d838c86
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/5199
Reviewed-by: Tony Georgiev <tgeorgiev@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>