Fix failing upgrade test
- The test was failing with an assertion error when trying to get
credentials. The reason was that the creation of the content used to
happen before admiral is fully started. Managed to reproduce the issue
locally by decreasing the time. It looks like that in jenkins the run is
slower. Since we do not have health check in the branch increasing the
time to wait before initializing the content
- Enabled again the check for the host since it looks like that the
problem was the same.

Change-Id: I131e6c898c67caf73cfdf42aeb1f3469fd2de92f
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/5560
Reviewed-by: Sergio Sanchez <sanchezs@vmware.com>
Bellevue-Verified: e_vcoauto_glob_1 <e_vcoauto_glob_1@vmware.com>
CS-Verified: e_vcoauto_glob_1 <e_vcoauto_glob_1@vmware.com>