Polished kubernetes models conversion.
Added conversion for cpu shares and tests to confirm the implementation.

Exported entity-specific methods in entity-to-entity converter class to
achieve better reusability and also KubernetesUtil is becoming too large
and hard to maintain. The idea is KubernetesUtil class to provide only
specific util methods to convert high-level objects. For example:
fromCompositeTemplateToPod or fromCompositeTemplateToDeployment and so
on.

Change-Id: If7bdc1950e33132bc81578d2ea96d263f91edeef
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/5572
Reviewed-by: Sergio Sanchez <sanchezs@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>