Fix upgrade integration test
The problem was that right after creating the entities in the old
admiral version the admiral container was removed. In some cases the
remove was executing while the data is being stored in lucene. This was
breaking the test and after upgrade the data was not availabale

Change-Id: I8c9a2b780a3b23dcaf8aece9eedbf5240fe35040
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/5608
Reviewed-by: Sergio Sanchez <sanchezs@vmware.com>
Bellevue-Verified: e_vcoauto_glob_1 <e_vcoauto_glob_1@vmware.com>
CS-Verified: e_vcoauto_glob_1 <e_vcoauto_glob_1@vmware.com>