Attempt to speedup maven build
I have been observing that most of the time is spent while
tests wait for the default services to be up and running. This
is because once we call registerForServiceAvailability
if the service is not available, we will check again after the
maintanance defined timeout - 250 ms. If we have 1000
of tests, this may lead to a total delay of 250 seconds.

Change-Id: I867a056703146089604edb4a721953e6d885164f
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/5712
Bellevue-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Peter Mitrov <pmitrov@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>