[VBV-991] Enable TemplateSearchServiceTest
- the problem was that the indexing of the "name" property
of container description state is set to be case insensitive.
This means that if the we create a container desc state with
name "tesT", it will set the name to "tesT" but the indexing
will be "test", so in the query filter the "name" property
have to be toLowerCase.

Change-Id: I51569a4d4090c32fdb5ade4a93965e2e99c33267
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/5842
Upgrade-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Tony Georgiev <tgeorgiev@vmware.com>