Modify the container state before redeploying the system container.
When upgrade the system container with newer version we take properties from its previous state.
In our case we use information from this state  as parameters when creating the new container. One
of this properties is container ip which causes the container to stuck in created state.
Here we modify the state before we use it for the deployment.

Change-Id: Ic6e10da01569e3697b2e9b3e413502488c199ada
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/5913
Upgrade-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Lazarin Lazarov <llazarov@vmware.com>