[UI] Fix selecting a network mode for containers
Due to recent changes of the Admiral look and feel, we were not able to
set the network mode for a given container. The problem was that the
network mode dropdown menu lost its "form-control" style, but it is used
in ContainerDefinitionForm.getRawInput() to select the dropdown component
from the DOM and then retrieve the selected option.

Change-Id: If3e7b6f5952243537ca4375b8659fa65c880d3dd
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/5984
Reviewed-by: Tony Georgiev <tgeorgiev@vmware.com>
Upgrade-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>