Fix ReservationTaskServiceTest.testReservationTaskLifeCycleWithNoGroup
The problem was the way placement is selected. When a request was made with
no group (i.e. tenantLinks == null) it would have eventually query all placement zones
so it would even pick a placement zone with any group.

That was causing the test to fail, and since only the first of both placements was picked
the chance to fail this test was 50%

Change-Id: I4111d0e30d47cb2e40fe378d3aa180cd43f8067a
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/6021
Reviewed-by: Rostislav Georgiev <rgeorgiev@vmware.com>
Upgrade-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>