VBV-1026 - Resolve deserialization issues from 7.3 for entities stored in 7.2
In the latest xenon there is a system property introduced and when
passed "-Dxenon.kryo.handleBuiltInCollections=false" it solves the
problem. However with with the property we could have issues with kryo
when using immutable collections (as it was with xenon <=1.1.1). The
problem here was that during provisioning we had NPE from kryo
- Changed the initialization of the ports in cpntainer state for
Arrays.asList() to new ArrayList

Change-Id: I507edd088b4b693bfbb32fea5783c25a40bae123
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/6181
Upgrade-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Lazarin Lazarov <llazarov@vmware.com>