Fix sporadic failures in ContainerRemovalTaskService
Some integration tests are failing in ContainerRemovalTaskService because,
while a container is removed, a data collection can start, mark the container
as missing and delete it. If this happens, ContainerRemovalTaskService fails
because it can't find the container in removeResources. In this case, the task
should not fail, just continue and finish with success (because both the container
and container state were already removed).

Change-Id: Iade36dc2ffaea0961f75698608f65b6a8e29c939
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/6385
Upgrade-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Georgi Muleshkov <gmuleshkov@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>