VBV-1012 Remove transient volumes state if retired
Transient volumes are discovered by data collection, but when the container is
removed, the corresponding volume is removed from the docker host, but the
state is not removed from the index.

Introduce a counter to track the number of failed data collections for every
such volume. Based on that counter, delete the volume state when necessary.

Change-Id: I1c0bb811e11530caa535025457d1ba69ea9e8807
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/6514
Upgrade-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Jose Dillet <jdillet@vmware.com>