Temporary backing off placement zone refresh on compute change
PZ refresh on a compute change is expensive and currently there
are a lot of such changes (especially during endpoint enumeration
or massive containter creation, etc.).
Temporary introducing a pause of 10 seconds after such a refresh
to make sure we don't over-stress the system. If a change happens
during the pause, a refresh is automatically triggered once the
pause is over.

Longer-term the actual refresh has to be optimized.

Also disallowed IDEMPOTENT_POST for periodic tasks. We want a POST
to fail if a task with the same self link is already running.

Change-Id: I53f9e02f7b922a2a1240d545e3028194e0e42073
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/6748
Upgrade-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Peter Mitrov <pmitrov@vmware.com>