VBV-1045: Improper request tracking when creating volume with non-existent plugin
When a request fails and moves to RESERVATION_CLEANUP and RESERVATION_CLEANED_UP
subtages in RequestBrokerState, the request status is marked as finished with
100% progress, but after that, it moves to ERROR, so the task is marked as
FAILED. From the UI, this shows the task green for a few seconds, and then
as red, which is very confusing. This was seen when creating a volume from
the CRUD UI with an invalid driver, but it can happen in all request types.

The root problem is that all tracked tasks are 100% completed, but there is
still some work to do in RequestBrokerService before reaching ERROR state.
To avoid having all the tracked tasks completed, the fix adds the
RequestBrokerService as tracked task.

Also, the tracked tasks should include the provisioning tasks when request is
not an allocation only request.

Change-Id: Id766e83c095c77263886dc687ee82a40ab7e0a51
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/7262
Reviewed-by: Sergio Sanchez <sanchezs@vmware.com>
Compute-Verified: jenkins <jenkins@vmware.com>
Upgrade-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Tony Georgiev <tgeorgiev@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>