Fetch logs implemented in KubernetesAdapterService
Implemented FETCH_LOGS operation in the kubernetes adapter. It is only
for pods, and it will create LogServiceState per container in the pod.

Added PodLogService which is similar to ContainerLogService and it
exposes fetching logs functionality. The difference is that
PodLogService will return map for the specified pod, where the key is
the container name and the value is the LogServiceState object

Added tests for both adapter and PodLogService.

Change-Id: Ibf6d37e59dfa03fa5618e37b6716a6906bbca7d4
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/7391
Upgrade-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Tony Georgiev <tgeorgiev@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Compute-Verified: jenkins <jenkins@vmware.com>