VBV-1108: When deploying WordPress_with_MySQL_volume.yaml from Catalog, it fails
This test uncovered an issue when there are several hosts, and we provision an
application with a clustered container connected to local volumes. When there
is at least one local volume, all the containers connected to it must be placed
on the same host. However, if the first container to be deployed was clustered,
the containers were spread amongst different hosts.

The fix introduces a filter, which runs after all other filters has passed. If
the container is clustered and with local volumes, if there is one more than
one host, it selects one host.

Change-Id: I8b09e8986f3eab649cfc047408ec58ccf44c97f6
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/7841
Compute-Verified: jenkins <jenkins@vmware.com>
Upgrade-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Miroslav Shipkovenski <mshipkovenski@vmware.com>