VBV-1023 - [Upgrade] Admiral agent is not upgraded automatically
The issue was reproducible on out jenkins workers. The reason was that
for some reason the image of the old system container was deleted.
During the data collection we compare the version from the image but
when the image is missing we get an id for version for example
"test:22335caac996". In this case the comparision of the version fail
and we used to return 0
- Changed the comparison to return -1 when the version is not in an
expected format. Like this if we cannot detect the version the system
container will be reprovisioned.
- Added a test for the case above

Change-Id: Iad6f3812c1bfb48f32de3cf298deb5bf6b37e8bf
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/7838
Compute-Verified: jenkins <jenkins@vmware.com>
Upgrade-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Jose Dillet <jdillet@vmware.com>