Bumping up xenon to 1.4.1
- Updated the version of xenon
- Changed the migration service not to fail if a migration task fails
because a factory is not available on the source. This is needed because
the behavior of a change in the behavior in the migration task service
in the new xenon.

Change-Id: I2570e16e1a28ade01b2ab3f0be8edc87a4b9ed47
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/8162
Reviewed-by: Peter Mitrov <pmitrov@vmware.com>
Compute-Verified: jenkins <jenkins@vmware.com>
Upgrade-Verified: jenkins <jenkins@vmware.com>
Closures-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: George P Dimitrov <dimitrovg@vmware.com>