VBV-1105: imageReference can not be set in templates
ContainerDescription.imageReference could not be set in yaml templates because
the property was ignored, so it used to be imported as customProperty.
This fix allows to set imageReference in yaml templates. Additionally, it allows
to use a file reference (file:///path_to_file). In case the scheme is file,
DockerAdapterService loads the file without downloading it to a temporary
file. The docker compose format is not affected by this change.

Finally, the fix adds a unit test for importing templates with imageReference,
and removed and old workaround for an issue loading a file which was fixed
long ago.

Change-Id: I72a44ce24cff3d1d8ff1a462a5998ef1a44d1046
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/7945
Compute-Verified: jenkins <jenkins@vmware.com>
Upgrade-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Tony Georgiev <tgeorgiev@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>