[VBV-1123] Fix intermittently failing request in cluster.
com.vmware.admiral.request.cluster.RequestInClusterNodesTest.testRequestLifeCycleInClusterWhenOneNodeDown is failing when validating the request. The reason for this is that the test assert the progress of RequestStatus is equal to 100. The update of progress status and patching of task to FINISHED stage is happening asynchronously, thats why there is chance the task to complete with progress < 100. In this change instead of asserting the progress is equal to 100, the test will poll until the progress is patched to 100.

Change-Id: Ic5defd7e39835f2da9fbfac375162dee928f32af
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/7980
Reviewed-by: Zahari Ivanov <zaharii@vmware.com>
Compute-Verified: jenkins <jenkins@vmware.com>
Upgrade-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>