Re-adding a host with wrong credentials passes
Credentials are not checked if the host is deleted and then re-added.
The reason is the xenon uses a channel pool for its connections and if
there is an open channel to the host we just deleted, when we re-add it,
the channel is reused, thus not verifying the credentials.

Because the channel is selected based on context and the context
identifyier is built from an operation tag, the remote host and port, a
simple solution is to set a tag for the operation, unique for the
credentials we varify.

Change-Id: I22bfc2c39def15312e450f68444cf1a8d142fc34
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/8331
Compute-Verified: jenkins <jenkins@vmware.com>
Upgrade-Verified: jenkins <jenkins@vmware.com>
Closures-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Sergio Sanchez <sanchezs@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>