VBV-1141 - [CAFE Catalog] Can't destroy application if you create a template from any of its containers
The problem was on how the clone composite description is used for the
"Create Template from container" operation. When we clone a composite
description it becomes the parent of the cloned one, but in case of
having an already running container, doing that and just changing the
container parent was leaving some outdated data in the placements...
and the remove container operation failed (the request) despite of
deleting the actual resources.

- Extended the clone composite description service to:
  - Do the proper patches of the child descriptions when necessary.
  - Allow creating a clone and mark it as the parent description at
  once (the reverse parameter). Doing the clone in this way for the
  "Create Template from container" operation solves the issue.
- Added some tests.
- Simplified the logic on the UI, instead of:
1. create a new composite description
2. clone it
3. patch the selected container description
now just 2 calls are needed:
1. create a new composite description
2. clone it
- Added the logic in the UI to allow creating a new template from a
container if the previous one was removed.

Change-Id: I381e841643f5f15e54b9f5bb07276ea53ca34c83
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/8576
Compute-Verified: jenkins <jenkins@vmware.com>
Upgrade-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
Closures-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Tony Georgiev <tgeorgiev@vmware.com>