Fix failing migration task because of /config/event-topic
The problem was that this is a default service that is started on every
node. Because of this the migration service tries to migrate it and it
fails with "service already started".
- Exclude the service from the list of services to migrate.

Change-Id: Ib54260eb86920cff2e9a3c8d1e5a0ab7274ba9ff
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/8791
Reviewed-by: Zahari Ivanov <zaharii@vmware.com>
Compute-Verified: jenkins <jenkins@vmware.com>
Closures-Verified: jenkins <jenkins@vmware.com>
Upgrade-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>