English is not added to the supported locales
In xenon we determine the supported locales by the extensions of the
file bundles. l10n tool doesn't support bundles without a default (no
extension) and we remove the extension for English, which has broken the
localization as it is now not discovered. Adding a file explicitly for
English, because this is the quickest way to fix this and I think it
should go into the release. We should discuss whether to change the way
we discover locales in xenon.

Change-Id: I7faef00b76098129e9d618cd6655cdc2d3159465
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/9024
Compute-Verified: jenkins <jenkins@vmware.com>
Closures-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Upgrade-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Elena Ivanova <eivanova@vmware.com>