VBV-1188 - In compute state the adapterReference should not be the full uri
The problem was that we in the compute state foe
adapterManagementReference we set the full uri starting with the host
and the port. With a single node this is not a problem. However in
cluster if the node that the references points to is down the data
collection does not work.
- Added check for the exception we get and in case it is
ConnectException change the reference.

Change-Id: Ib1c8813d24554727dcced0688cbff13f29fb7a6f
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/9181
Upgrade-Verified: jenkins <jenkins@vmware.com>
Compute-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Closures-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Antonio Filipov <afilipov@vmware.com>
Reviewed-by: Peter Mitrov <pmitrov@vmware.com>