[VBV-1193] Deallocation reqeust from container service
- the destroy operation from the composition is on 2 steps.
DESTROY which destroys the host and deallocates all the resources
from admiral and send back a callback to composition. The composition
send DEALLOCATION reqeust to us, and we again try to deallocate
resources which are already deallocated.

- the problem is that we already deallocated the resources in the
DESTROY phase, and there is no need of the DEALLOCATION phase.
But we don't control the composition requests and we somehow need
to differ the DEALLOCATION request.

- with the current implementation if the PATCH request
fails on the DESTROY phase the task will complete which
will lead to left resources on admiral side, but will
delete the resources from the composition because we
send COMPLETE request to it.

- added a check if the request is DEALLOCATION, and if it is
to COMPLETE the operation.

Change-Id: I1af970433df1a4a98386eb5f797b1cd019bd9022
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/9325
Compute-Verified: jenkins <jenkins@vmware.com>
Closures-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
Upgrade-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Rostislav Georgiev <rgeorgiev@vmware.com>