[VBV-1181] Avoid sending invalid request to ContainerImageService.
When request is handled from TemplateSearchService with query paramter
similar to"q=*", every leading asterisks will be removed, because
the registry will reject them. In this case after asterisks removal,
the query will be empty and when this is sent to ContainerImageService,
it will be invalid because there is validation that the query is not
empty, so instead of sending this kind of request, give the callback empty result.

Change-Id: I68c69ffe89a32fad94ac25cd3164c49abc897d1a
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/9483
Compute-Verified: jenkins <jenkins@vmware.com>
Closures-Verified: jenkins <jenkins@vmware.com>
Upgrade-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Tony Georgiev <tgeorgiev@vmware.com>